+++ src/frontend/apps/drive/src/features/explorer/components/filters/ExplorerFilterType.tsx
@@ -1,0 +1,52 @@
+import { Filter, FilterOption } from "@gouvfr-lasuite/ui-kit";
+import { useMemo } from "react";
+import { useTranslation } from "react-i18next";
+import { Key } from "react-aria-components";
+import folderIcon from "@/assets/folder/folder.svg";
+import mimeOther from "@/assets/files/icons/mime-other.svg";
+import { ItemType } from "@/features/drivers/types";
+import { getResetOption } from "./filterUtils";
+
+export const ExplorerFilterType = (props: {
+  value: ItemType | null;
+  onChange: (value: Key | null) => void;
+}) => {
+  const { t } = useTranslation();
+
+  const typeOptions: FilterOption[] = useMemo(
+    () => [
+      {
+        label: t("explorer.filters.type.options.folder"),
+        value: "folder",
+        render: () => (
+          <div className="explorer__filters__item">
+            <img src={folderIcon.src} alt="" width="24" height="24" />
+            {t("explorer.filters.type.options.folder")}
+          </div>
+        ),
+      },
+      {
+        label: t("explorer.filters.type.options.file"),
+        render: () => (
+          <div className="explorer__filters__item">
+            <img src={mimeOther.src} alt="" width="24" height="24" />
+            {t("explorer.filters.type.options.file")}
+          </div>
+        ),
+        value: "file",
+        showSeparator: true,
+      },
+      getResetOption(t),
+    ],
+    [t],
+  );
+
+  return (
+    <Filter
+      label={t("explorer.filters.type.label")}
+      options={typeOptions}
+      selectedKey={props.value ?? null} // undefined would trigger "uncontrolled components become controlled" warning.
+      onSelectionChange={props.onChange}
+    />
+  );
+};
